refactor: 에피그램 상세 페이지 SSR로 변경
